requirements.txt
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
PyYAML==3.12
